Factory default and instance settings loader. Move web app/ws and news_api app executable into the core.
