--- newsroom/web/__init__.py
@@ -1,1 +1,0 @@
-from .app import NewsroomWebApp, get_app  # noqa
